audited and updated core and observability modules
