+++ secureli/modules/observability/observability_services/logging.py
@@ -1,57 +1,12 @@
-import platform
-from datetime import datetime
-from enum import Enum
 from pathlib import Path
 from typing import Optional
-from uuid import uuid4
 
-import pydantic
-from secureli.modules.shared.models.config import HookConfiguration
-from secureli.modules.shared.models.logging import LogAction
-
+import secureli.modules.shared.models.logging as LoggingModels
 import secureli.repositories.secureli_config as SecureliConfig
+
 from secureli.modules.language_analyzer import language_support
 from secureli.repositories.secureli_config import SecureliConfigRepository
 from secureli.modules.shared import utilities
-
-
-def generate_unique_id() -> str:
-    """
-    A unique identifier representing the log entry, including various
-    bits specific to the user and environment
-    """
-    origin_email_branch = f"{utilities.origin_url()}|{utilities.git_user_email()}|{utilities.current_branch_name()}"
-    return f"{uuid4()}|{origin_email_branch}"
-
-
-class LogStatus(str, Enum):
-    """Whether the entry represents a successful or failing entry"""
-
-    success = "SUCCESS"
-    failure = "FAILURE"
-
-
-class LogFailure(pydantic.BaseModel):
-    """An extendable structure for log failures"""
-
-    details: str
-
-
-class LogEntry(pydantic.BaseModel):
-    """A distinct entry in the log captured following actions like scan and init"""
-
-    id: str = generate_unique_id()
-    timestamp: datetime = datetime.utcnow()
-    username: str = utilities.git_user_email()
-    machineid: str = platform.uname().node
-    secureli_version: str = utilities.secureli_version()
-    languages: Optional[list[str]]
-    status: LogStatus
-    action: LogAction
-    hook_config: Optional[HookConfiguration]
-    failure: Optional[LogFailure] = None
-    total_failure_count: Optional[int]
-    failure_count_details: Optional[object]
 
 
 class LoggingService:
@@ -65,7 +20,7 @@ class LoggingService:
         self.language_support = language_support
         self.secureli_config = secureli_config
 
-    def success(self, action: LogAction) -> LogEntry:
+    def success(self, action: LoggingModels.LogAction) -> LoggingModels.LogEntry:
         """
         Capture that a successful conclusion has been reached for an action
         :param action: The action that succeeded
@@ -76,8 +31,8 @@ class LoggingService:
             if secureli_config.languages
             else None
         )
-        log_entry = LogEntry(
-            status=LogStatus.success,
+        log_entry = LoggingModels.LogEntry(
+            status=LoggingModels.LogStatus.success,
             action=action,
             hook_config=hook_config,
             languages=secureli_config.languages if secureli_config.languages else None,
@@ -88,11 +43,11 @@ class LoggingService:
 
     def failure(
         self,
-        action: LogAction,
+        action: LoggingModels.LogAction,
         details: str,
         total_failure_count: Optional[int] = None,
         individual_failure_count: Optional[object] = None,
-    ) -> LogEntry:
+    ) -> LoggingModels.LogEntry:
         """
         Capture a failure against an action, with details
         :param action: The action that failed
@@ -106,10 +61,10 @@ class LoggingService:
             if not secureli_config.languages
             else self.language_support.get_configuration(secureli_config.languages)
         )
-        log_entry = LogEntry(
-            status=LogStatus.failure,
+        log_entry = LoggingModels.LogEntry(
+            status=LoggingModels.LogStatus.failure,
             action=action,
-            failure=LogFailure(
+            failure=LoggingModels.LogFailure(
                 details=details,
             ),
             total_failure_count=total_failure_count,
@@ -121,7 +76,7 @@ class LoggingService:
 
         return log_entry
 
-    def _log(self, log_entry: LogEntry):
+    def _log(self, log_entry: LoggingModels.LogEntry):
         """Commit a log entry to the branch log file"""
         log_folder_path = Path(SecureliConfig.FOLDER_PATH / ".secureli/logs")
         path_to_log = log_folder_path / f"{utilities.current_branch_name()}"
