+++ secureli/modules/shared/models/update.py
@@ -1,0 +1,12 @@
+from typing import Optional
+
+import pydantic
+
+
+class UpdateResult(pydantic.BaseModel):
+    """
+    The results of calling scan_repo
+    """
+
+    successful: bool
+    output: Optional[str] = None
